modificaciones vendedor,excepcion,archivo,menu,tienda
Se quito el id del vendendor
Se cambio el login de id a dni
Se modifico la excepcion id incorrecto a dni incorrecto
Se realizaron funciones en el menu para la facturacion y el logeo
Se añadieron funciones en tienda informatica  para login y vendedores
Se modifico en la clase de archivos que cuando se vuelquen las facturas del archivo a la tienda la clave sea un integer

+++ src/manejoArchivo/ArchivoHerramientas.java
@@ -183,7 +183,7 @@ public class ArchivoHerramientas<K> {
 			
 			while((facturaAux=(Factura)facturaArchi.readObject())!=null)
 			{
-				contenedoraFacturas.agregarObjetoColeccion(facturaAux.getId(), facturaAux);
+				contenedoraFacturas.agregarObjetoColeccion((Integer)facturaAux.getId(), facturaAux);
 			}
 			facturaArchi.close();
 		} catch (FileNotFoundException e) {
